requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
opencv-python
pillow
matplotlib
seaborn
requests
tqdm
geopandas
python +-dotenv
